rusqlite::auto_extension

Function init_auto_extension

pub unsafe fn init_auto_extension(
    db: *mut sqlite3,
    pz_err_msg: *mut *mut c_char,
    ax: AutoExtension,
) -> c_int

Bridge between RawAutoExtension and AutoExtension

Safety

Opening a database from an auto-extension handler will lead to an endless recursion of the auto-handler triggering itself indirectly for each newly-opened database.

Results are undefined if the given db is closed by an auto-extension.

The list of auto-extensions should not be manipulated from an auto-extension.