rusqlite::auto_extension

Function register_auto_extension

pub unsafe fn register_auto_extension(ax: RawAutoExtension) -> Result<()>

Register au auto-extension

Safety

Opening a database from an auto-extension handler will lead to an endless recursion of the auto-handler triggering itself indirectly for each newly-opened database.

Results are undefined if the given db is closed by an auto-extension.

The list of auto-extensions should not be manipulated from an auto-extension.