Struct rusqlite::blob::ZeroBlob

pub struct ZeroBlob(pub i32);

BLOB of length N that is filled with zeroes.

Zeroblobs are intended to serve as placeholders for BLOBs whose content is later written using incremental BLOB I/O routines.

A negative value for the zeroblob results in a zero-length BLOB.