Trait rusqlite::functions::FromValue

pub trait FromValue: Sized {
    unsafe fn parameter_value(v: *mut sqlite3_value) -> Result<Self>;

    unsafe fn parameter_has_valid_sqlite_type(_: *mut sqlite3_value) -> bool { ... }
}

A trait for types that can be created from a SQLite function parameter value.

Required Methods

unsafe fn parameter_value(v: *mut sqlite3_value) -> Result<Self>

Provided Methods

unsafe fn parameter_has_valid_sqlite_type(_: *mut sqlite3_value) -> bool

FromValue types can implement this method and use sqlite3_value_type to check that the type reported by SQLite matches a type suitable for Self. This method is used by Context::get to confirm that the parameter contains a valid type before attempting to retrieve the value.