Enum rusqlite::limits::Limit

#[repr(C)]

pub enum Limit {
    SQLITE_LIMIT_LENGTH,
    SQLITE_LIMIT_SQL_LENGTH,
    SQLITE_LIMIT_COLUMN,
    SQLITE_LIMIT_EXPR_DEPTH,
    SQLITE_LIMIT_COMPOUND_SELECT,
    SQLITE_LIMIT_VDBE_OP,
    SQLITE_LIMIT_FUNCTION_ARG,
    SQLITE_LIMIT_ATTACHED,
    SQLITE_LIMIT_LIKE_PATTERN_LENGTH,
    SQLITE_LIMIT_VARIABLE_NUMBER,
    SQLITE_LIMIT_TRIGGER_DEPTH,
    SQLITE_LIMIT_WORKER_THREADS,
}

Run-Time Limit Categories

Variants
SQLITE_LIMIT_LENGTH

The maximum size of any string or BLOB or table row, in bytes.

SQLITE_LIMIT_SQL_LENGTH

The maximum length of an SQL statement, in bytes.

SQLITE_LIMIT_COLUMN

The maximum number of columns in a table definition or in the result set of a SELECT or the maximum number of columns in an index or in an ORDER BY or GROUP BY clause.

SQLITE_LIMIT_EXPR_DEPTH

The maximum depth of the parse tree on any expression.

SQLITE_LIMIT_COMPOUND_SELECT

The maximum number of terms in a compound SELECT statement.

SQLITE_LIMIT_VDBE_OP

The maximum number of instructions in a virtual machine program used to implement an SQL statement.

SQLITE_LIMIT_FUNCTION_ARG

The maximum number of arguments on a function.

SQLITE_LIMIT_ATTACHED

The maximum number of attached databases.

SQLITE_LIMIT_LIKE_PATTERN_LENGTH

The maximum length of the pattern argument to the LIKE or GLOB operators.

SQLITE_LIMIT_VARIABLE_NUMBER

The maximum index number of any parameter in an SQL statement.

SQLITE_LIMIT_TRIGGER_DEPTH

The maximum depth of recursion for triggers.

SQLITE_LIMIT_WORKER_THREADS

The maximum number of auxiliary worker threads that a single prepared statement may start.