rusqlite

Module row

Modules

sealed 🔒

Macros

tuple_try_from_row 🔒

tuples_try_from_row 🔒

Structs

AndThenRows

An iterator over the mapped resulting rows of a query, with an Error type unifying with Error.

Map

F is used to transform the streaming iterator into a fallible iterator.

MappedRows

An iterator over the mapped resulting rows of a query.

Row

A single result row of a query.

Rows

A handle (lazy fallible streaming iterator) for the resulting rows of a query.

Traits

RowIndex

A trait implemented by types that can index into columns of a row.