rusqlite::row

Trait RowIndex

pub trait RowIndex: Sealed {
    // Required method
    fn idx(&self, stmt: &Statement<'_>) -> Result<usize>;
}

A trait implemented by types that can index into columns of a row.

It is only implemented for usize and &str.

Required Methods

fn idx(&self, stmt: &Statement<'_>) -> Result<usize>

Returns the index of the appropriate column, or Error if no such column exists.