Struct rusqlite::CachedStatement

pub struct CachedStatement<'conn> { /* fields omitted */ }

Cacheable statement.

Statement will return automatically to the cache by default. If you want the statement to be discarded, call discard() on it.

Methods

impl<'conn> CachedStatement<'conn>

fn discard(self)

Discard the statement, preventing it from being returned to its Connection's collection of cached statements.