rusqlite

Struct Connection

pub struct Connection {
    pub(crate) db: RefCell<InnerConnection>,
    pub(crate) cache: StatementCache,
    pub(crate) transaction_behavior: TransactionBehavior,
}

A connection to a SQLite database.

Fields
db: RefCell<InnerConnection>cache: StatementCachetransaction_behavior: TransactionBehavior
Implementations

impl Connection

pub fn busy_timeout(&self, timeout: Duration) -> Result<()>

Set a busy handler that sleeps for a specified amount of time when a table is locked. The handler will sleep multiple times until at least “ms” milliseconds of sleeping have accumulated.

Calling this routine with an argument equal to zero turns off all busy handlers.

There can only be a single busy handler for a particular database connection at any given moment. If another busy handler was defined (using busy_handler) prior to calling this routine, that other busy handler is cleared.

Newly created connections currently have a default busy timeout of 5000ms, but this may be subject to change.

pub fn busy_handler(&self, callback: Option<fn(i32) -> bool>) -> Result<()>

Register a callback to handle SQLITE_BUSY errors.

If the busy callback is None, then SQLITE_BUSY is returned immediately upon encountering the lock. The argument to the busy handler callback is the number of times that the busy handler has been invoked previously for the same locking event. If the busy callback returns false, then no additional attempts are made to access the database and SQLITE_BUSY is returned to the application. If the callback returns true, then another attempt is made to access the database and the cycle repeats.

There can only be a single busy handler defined for each database connection. Setting a new busy handler clears any previously set handler. Note that calling busy_timeout() or evaluating PRAGMA busy_timeout=N will change the busy handler and thus clear any previously set busy handler.

Newly created connections default to a busy_timeout() handler with a timeout of 5000ms, although this is subject to change.

impl Connection

pub fn prepare_cached(&self, sql: &str) -> Result<CachedStatement<'_>>

Prepare a SQL statement for execution, returning a previously prepared (but not currently in-use) statement if one is available. The returned statement will be cached for reuse by future calls to prepare_cached once it is dropped.

fn insert_new_people(conn: &Connection) -> Result<()> {
    {
        let mut stmt = conn.prepare_cached("INSERT INTO People (name) VALUES (?1)")?;
        stmt.execute(["Joe Smith"])?;
    }
    {
        // This will return the same underlying SQLite statement handle without
        // having to prepare it again.
        let mut stmt = conn.prepare_cached("INSERT INTO People (name) VALUES (?1)")?;
        stmt.execute(["Bob Jones"])?;
    }
    Ok(())
}

Failure

Will return Err if sql cannot be converted to a C-compatible string or if the underlying SQLite call fails.

pub fn set_prepared_statement_cache_capacity(&self, capacity: usize)

Set the maximum number of cached prepared statements this connection will hold. By default, a connection will hold a relatively small number of cached statements. If you need more, or know that you will not use cached statements, you can set the capacity manually using this method.

pub fn flush_prepared_statement_cache(&self)

Remove/finalize all prepared statements currently in the cache.

impl Connection

pub fn column_exists<N: Name>( &self, db_name: Option<N>, table_name: N, column_name: N, ) -> Result<bool>

Check if table_name.column_name exists.

db_name is main, temp, the name in ATTACH, or None to search all databases.

pub fn table_exists<N: Name>( &self, db_name: Option<N>, table_name: N, ) -> Result<bool>

Check if table_name exists.

db_name is main, temp, the name in ATTACH, or None to search all databases.

pub fn column_metadata<N: Name>( &self, db_name: Option<N>, table_name: N, column_name: N, ) -> Result<(Option<&CStr>, Option<&CStr>, bool, bool, bool)>

Extract metadata of column at specified index

Returns:

declared data type

name of default collation sequence

True if column has a NOT NULL constraint

True if column is part of the PRIMARY KEY

True if column is AUTOINCREMENT

fn exists<N: Name>( &self, db_name: Option<N>, table_name: N, column_name: Option<N>, ) -> Result<bool>

impl Connection

pub fn db_config(&self, config: DbConfig) -> Result<bool>

Returns the current value of a config.

SQLITE_DBCONFIG_ENABLE_FKEY: return false or true to indicate whether FK enforcement is off or on

SQLITE_DBCONFIG_ENABLE_TRIGGER: return false or true to indicate whether triggers are disabled or enabled

SQLITE_DBCONFIG_ENABLE_FTS3_TOKENIZER: return false or true to indicate whether fts3_tokenizer are disabled or enabled

SQLITE_DBCONFIG_NO_CKPT_ON_CLOSE: return false to indicate checkpoints-on-close are not disabled or true if they are

SQLITE_DBCONFIG_ENABLE_QPSG: return false or true to indicate whether the QPSG is disabled or enabled

SQLITE_DBCONFIG_TRIGGER_EQP: return false to indicate output-for-trigger are not disabled or true if it is

pub fn set_db_config(&self, config: DbConfig, new_val: bool) -> Result<bool>

Make configuration changes to a database connection

SQLITE_DBCONFIG_ENABLE_FKEY: false to disable FK enforcement, true to enable FK enforcement

SQLITE_DBCONFIG_ENABLE_TRIGGER: false to disable triggers, true to enable triggers

SQLITE_DBCONFIG_ENABLE_FTS3_TOKENIZER: false to disable fts3_tokenizer(), true to enable fts3_tokenizer()

SQLITE_DBCONFIG_NO_CKPT_ON_CLOSE: false (the default) to enable checkpoints-on-close, true to disable them

SQLITE_DBCONFIG_ENABLE_QPSG: false to disable the QPSG, true to enable QPSG

SQLITE_DBCONFIG_TRIGGER_EQP: false to disable output for trigger programs, true to enable it

impl Connection

pub fn pragma_query_value<T, F>( &self, schema_name: Option<&str>, pragma_name: &str, f: F, ) -> Result<T>
where F: FnOnce(&Row<'_>) -> Result<T>,

Query the current value of pragma_name.

Some pragmas will return multiple rows/values which cannot be retrieved with this method.

Prefer PRAGMA function introduced in SQLite 3.20: SELECT user_version FROM pragma_user_version;

pub fn pragma_query<F>( &self, schema_name: Option<&str>, pragma_name: &str, f: F, ) -> Result<()>
where F: FnMut(&Row<'_>) -> Result<()>,

Query the current rows/values of pragma_name.

Prefer PRAGMA function introduced in SQLite 3.20: SELECT * FROM pragma_collation_list;

pub fn pragma<F, V>( &self, schema_name: Option<&str>, pragma_name: &str, pragma_value: V, f: F, ) -> Result<()>
where F: FnMut(&Row<'_>) -> Result<()>, V: ToSql,

Query the current value(s) of pragma_name associated to pragma_value.

This method can be used with query-only pragmas which need an argument (e.g. table_info('one_tbl')) or pragmas which returns value(s) (e.g. integrity_check).

Prefer PRAGMA function introduced in SQLite 3.20: SELECT * FROM pragma_table_info(?1);

pub fn pragma_update<V>( &self, schema_name: Option<&str>, pragma_name: &str, pragma_value: V, ) -> Result<()>
where V: ToSql,

Set a new value to pragma_name.

Some pragmas will return the updated value which cannot be retrieved with this method.

pub fn pragma_update_and_check<F, T, V>( &self, schema_name: Option<&str>, pragma_name: &str, pragma_value: V, f: F, ) -> Result<T>
where F: FnOnce(&Row<'_>) -> Result<T>, V: ToSql,

Set a new value to pragma_name and return the updated value.

Only few pragmas automatically return the updated value.

impl Connection

pub fn transaction(&mut self) -> Result<Transaction<'_>>

Begin a new transaction with the default behavior (DEFERRED).

The transaction defaults to rolling back when it is dropped. If you want the transaction to commit, you must call commit or set_drop_behavior(DropBehavior::Commit).

Example

fn perform_queries(conn: &mut Connection) -> Result<()> {
    let tx = conn.transaction()?;

    do_queries_part_1(&tx)?; // tx causes rollback if this fails
    do_queries_part_2(&tx)?; // tx causes rollback if this fails

    tx.commit()
}

Failure

Will return Err if the underlying SQLite call fails.

pub fn transaction_with_behavior( &mut self, behavior: TransactionBehavior, ) -> Result<Transaction<'_>>

Begin a new transaction with a specified behavior.

See transaction.

Failure

Will return Err if the underlying SQLite call fails.

pub fn unchecked_transaction(&self) -> Result<Transaction<'_>>

Begin a new transaction with the default behavior (DEFERRED).

Attempt to open a nested transaction will result in a SQLite error. Connection::transaction prevents this at compile time by taking &mut self, but Connection::unchecked_transaction() may be used to defer the checking until runtime.

See Connection::transaction and Transaction::new_unchecked (which can be used if the default transaction behavior is undesirable).

Example

fn perform_queries(conn: Rc<Connection>) -> Result<()> {
    let tx = conn.unchecked_transaction()?;

    do_queries_part_1(&tx)?; // tx causes rollback if this fails
    do_queries_part_2(&tx)?; // tx causes rollback if this fails

    tx.commit()
}

Failure

Will return Err if the underlying SQLite call fails. The specific error returned if transactions are nested is currently unspecified.

pub fn savepoint(&mut self) -> Result<Savepoint<'_>>

Begin a new savepoint with the default behavior (DEFERRED).

The savepoint defaults to rolling back when it is dropped. If you want the savepoint to commit, you must call commit or set_drop_behavior(DropBehavior::Commit).

Example

fn perform_queries(conn: &mut Connection) -> Result<()> {
    let sp = conn.savepoint()?;

    do_queries_part_1(&sp)?; // sp causes rollback if this fails
    do_queries_part_2(&sp)?; // sp causes rollback if this fails

    sp.commit()
}

Failure

Will return Err if the underlying SQLite call fails.

pub fn savepoint_with_name<T: Into<String>>( &mut self, name: T, ) -> Result<Savepoint<'_>>

Begin a new savepoint with a specified name.

See savepoint.

Failure

Will return Err if the underlying SQLite call fails.

pub fn transaction_state<N: Name>( &self, db_name: Option<N>, ) -> Result<TransactionState>

Determine the transaction state of a database

pub fn set_transaction_behavior(&mut self, behavior: TransactionBehavior)

Set the default transaction behavior for the connection.

Note

This will only apply to transactions initiated by transaction or unchecked_transaction.

Example

fn perform_queries(conn: &mut Connection) -> Result<()> {
    conn.set_transaction_behavior(TransactionBehavior::Immediate);

    let tx = conn.transaction()?;

    do_queries_part_1(&tx)?; // tx causes rollback if this fails
    do_queries_part_2(&tx)?; // tx causes rollback if this fails

    tx.commit()
}

impl Connection

pub fn open<P: AsRef<Path>>(path: P) -> Result<Self>

Open a new connection to a SQLite database. If a database does not exist at the path, one is created.

fn open_my_db() -> Result<()> {
    let path = "./my_db.db3";
    let db = Connection::open(path)?;
    // Use the database somehow...
    println!("{}", db.is_autocommit());
    Ok(())
}

Flags

Connection::open(path) is equivalent to using Connection::open_with_flags with the default OpenFlags. That is, it’s equivalent to:

Connection::open_with_flags(
    path,
    OpenFlags::SQLITE_OPEN_READ_WRITE
        | OpenFlags::SQLITE_OPEN_CREATE
        | OpenFlags::SQLITE_OPEN_URI
        | OpenFlags::SQLITE_OPEN_NO_MUTEX,
)

These flags have the following effects:

Open the database for both reading or writing.

Create the database if one does not exist at the path.

Allow the filename to be interpreted as a URI (see https://www.sqlite.org/uri.html#uri_filenames_in_sqlite for details).

Disables the use of a per-connection mutex.

Rusqlite enforces thread-safety at compile time, so additional locking is not needed and provides no benefit. (See the documentation on OpenFlags::SQLITE_OPEN_FULL_MUTEX for some additional discussion about this).

Most of these are also the default settings for the C API, although technically the default locking behavior is controlled by the flags used when compiling SQLite – rather than let it vary, we choose NO_MUTEX because it’s a fairly clearly the best choice for users of this library.

Failure

Will return Err if path cannot be converted to a C-compatible string or if the underlying SQLite open call fails.

WASM support

If you plan to use this connection type on the wasm32-unknown-unknown target please make sure to read the following notes:

The database is stored in memory by default.

Persistent VFS (Virtual File Systems) is optional, see https://github.com/Spxg/sqlite-wasm-rs for details

pub fn open_in_memory() -> Result<Self>

Open a new connection to an in-memory SQLite database.

Failure

Will return Err if the underlying SQLite open call fails.

pub fn open_with_flags<P: AsRef<Path>>( path: P, flags: OpenFlags, ) -> Result<Self>

Open a new connection to a SQLite database.

Database Connection for a description of valid flag combinations.

Failure

Will return Err if path cannot be converted to a C-compatible string or if the underlying SQLite open call fails.

pub fn open_with_flags_and_vfs<P: AsRef<Path>, V: Name>( path: P, flags: OpenFlags, vfs: V, ) -> Result<Self>

Open a new connection to a SQLite database using the specific flags and vfs name.

Database Connection for a description of valid flag combinations.

Failure

Will return Err if either path or vfs cannot be converted to a C-compatible string or if the underlying SQLite open call fails.

pub fn open_in_memory_with_flags(flags: OpenFlags) -> Result<Self>

Open a new connection to an in-memory SQLite database.

Database Connection for a description of valid flag combinations.

Failure

Will return Err if the underlying SQLite open call fails.

pub fn open_in_memory_with_flags_and_vfs<V: Name>( flags: OpenFlags, vfs: V, ) -> Result<Self>

Open a new connection to an in-memory SQLite database using the specific flags and vfs name.

Database Connection for a description of valid flag combinations.

Failure

Will return Err if vfs cannot be converted to a C-compatible string or if the underlying SQLite open call fails.

pub fn execute_batch(&self, sql: &str) -> Result<()>

Convenience method to run multiple SQL statements (that cannot take any parameters).

Example

fn create_tables(conn: &Connection) -> Result<()> {
    conn.execute_batch(
        "BEGIN;
         CREATE TABLE foo(x INTEGER);
         CREATE TABLE bar(y TEXT);
         COMMIT;",
    )
}

Failure

Will return Err if sql cannot be converted to a C-compatible string or if the underlying SQLite call fails.

pub fn execute<P: Params>(&self, sql: &str, params: P) -> Result<usize>

Convenience method to prepare and execute a single SQL statement.

On success, returns the number of rows that were changed or inserted or deleted (via sqlite3_changes).

Example

With positional params

fn update_rows(conn: &Connection) {
    match conn.execute("UPDATE foo SET bar = 'baz' WHERE qux = ?1", [1i32]) {
        Ok(updated) => println!("{} rows were updated", updated),
        Err(err) => println!("update failed: {}", err),
    }
}

With positional params of varying types

fn update_rows(conn: &Connection) {
    match conn.execute(
        "UPDATE foo SET bar = 'baz' WHERE qux = ?1 AND quux = ?2",
        params![1i32, 1.5f64],
    ) {
        Ok(updated) => println!("{} rows were updated", updated),
        Err(err) => println!("update failed: {}", err),
    }
}

With named params

fn insert(conn: &Connection) -> Result<usize> {
    conn.execute(
        "INSERT INTO test (name) VALUES (:name)",
        &[(":name", "one")],
    )
}

Failure

Will return Err if sql cannot be converted to a C-compatible string or if the underlying SQLite call fails.

pub fn path(&self) -> Option<&str>

Returns the path to the database file, if one exists and is known.

Returns Some("") for a temporary or in-memory database.

Note that in some cases PRAGMA database_list is likely to be more robust.

pub fn release_memory(&self) -> Result<()>

Attempts to free as much heap memory as possible from the database connection.

This calls sqlite3_db_release_memory.

pub fn last_insert_rowid(&self) -> i64

Get the SQLite rowid of the most recent successful INSERT.

Uses sqlite3_last_insert_rowid under the hood.

pub fn query_row<T, P, F>(&self, sql: &str, params: P, f: F) -> Result<T>
where P: Params, F: FnOnce(&Row<'_>) -> Result<T>,

Convenience method to execute a query that is expected to return a single row.

Example

fn preferred_locale(conn: &Connection) -> Result<String> {
    conn.query_row(
        "SELECT value FROM preferences WHERE name='locale'",
        [],
        |row| row.get(0),
    )
}

If the query returns more than one row, all rows except the first are ignored.

Returns Err(QueryReturnedNoRows) if no results are returned. If the query truly is optional, you can call .optional() on the result of this to get a Result<Option<T>>.

Failure

Will return Err if sql cannot be converted to a C-compatible string or if the underlying SQLite call fails.

pub fn query_one<T, P, F>(&self, sql: &str, params: P, f: F) -> Result<T>
where P: Params, F: FnOnce(&Row<'_>) -> Result<T>,

Convenience method to execute a query that is expected to return exactly one row.

Returns Err(QueryReturnedMoreThanOneRow) if the query returns more than one row.

Returns Err(QueryReturnedNoRows) if no results are returned. If the query truly is optional, you can call .optional() on the result of this to get a Result<Option<T>> (requires that the trait rusqlite::OptionalExtension is imported).

Failure

Will return Err if the underlying SQLite call fails.

pub fn query_row_and_then<T, E, P, F>( &self, sql: &str, params: P, f: F, ) -> Result<T, E>
where P: Params, F: FnOnce(&Row<'_>) -> Result<T, E>, E: From<Error>,

Convenience method to execute a query that is expected to return a single row, and execute a mapping via f on that returned row with the possibility of failure. The Result type of f must implement std::convert::From<Error>.

Example

fn preferred_locale(conn: &Connection) -> Result<String> {
    conn.query_row_and_then(
        "SELECT value FROM preferences WHERE name='locale'",
        [],
        |row| row.get(0),
    )
}

If the query returns more than one row, all rows except the first are ignored.

Failure

Will return Err if sql cannot be converted to a C-compatible string or if the underlying SQLite call fails.

pub fn prepare(&self, sql: &str) -> Result<Statement<'_>>

Prepare a SQL statement for execution.

Example

fn insert_new_people(conn: &Connection) -> Result<()> {
    let mut stmt = conn.prepare("INSERT INTO People (name) VALUES (?1)")?;
    stmt.execute(["Joe Smith"])?;
    stmt.execute(["Bob Jones"])?;
    Ok(())
}

Failure

Will return Err if sql cannot be converted to a C-compatible string or if the underlying SQLite call fails.

pub fn prepare_with_flags( &self, sql: &str, flags: PrepFlags, ) -> Result<Statement<'_>>

Prepare a SQL statement for execution.

Failure

Will return Err if sql cannot be converted to a C-compatible string or if the underlying SQLite call fails.

pub fn close(self) -> Result<(), (Self, Error)>

Close the SQLite connection.

This is functionally equivalent to the Drop implementation for Connection except that on failure, it returns an error and the connection itself (presumably so closing can be attempted again).

Failure

Will return Err if the underlying SQLite call fails.

pub unsafe fn handle(&self) -> *mut sqlite3

Get access to the underlying SQLite database connection handle.

Warning

You should not need to use this function. If you do need to, please open an issue on the rusqlite repository and describe your use case.

Safety

This function is unsafe because it gives you raw access to the SQLite connection, and what you do with it could impact the safety of this Connection.

pub unsafe fn from_handle(db: *mut sqlite3) -> Result<Self>

Create a Connection from a raw handle.

The underlying SQLite database connection handle will not be closed when the returned connection is dropped/closed.

Safety

This function is unsafe because improper use may impact the Connection.

pub unsafe fn from_handle_owned(db: *mut sqlite3) -> Result<Self>

Create a Connection from a raw owned handle.

The returned connection will attempt to close the inner connection when dropped/closed. This function should only be called on connections owned by the caller.

Safety

This function is unsafe because improper use may impact the Connection. In particular, it should only be called on connections created and owned by the caller, e.g. as a result of calling ffi::sqlite3_open().

pub fn get_interrupt_handle(&self) -> InterruptHandle

Get access to a handle that can be used to interrupt long-running queries from another thread.

pub(crate) fn decode_result(&self, code: c_int) -> Result<()>

pub fn changes(&self) -> u64

Return the number of rows modified, inserted or deleted by the most recently completed INSERT, UPDATE or DELETE statement on the database connection.

See https://www.sqlite.org/c3ref/changes.html

pub fn total_changes(&self) -> u64

Return the total number of rows modified, inserted or deleted by all completed INSERT, UPDATE or DELETE statements since the database connection was opened, including those executed as part of trigger programs.

See https://www.sqlite.org/c3ref/total_changes.html

pub fn is_autocommit(&self) -> bool

Test for auto-commit mode. Autocommit mode is on by default.

pub fn is_busy(&self) -> bool

Determine if all associated prepared statements have been reset.

pub fn cache_flush(&self) -> Result<()>

Flush caches to disk mid-transaction

pub fn is_readonly<N: Name>(&self, db_name: N) -> Result<bool>

Determine if a database is read-only

pub fn db_name(&self, index: usize) -> Result<String>

Return the schema name for a database connection

Failure

Return an Error::InvalidDatabaseIndex if index is out of range.

pub fn is_interrupted(&self) -> bool

Determine whether an interrupt is currently in effect