Struct rusqlite::OpenFlags

#[repr(C)]
pub struct OpenFlags(_);

Flags for opening SQLite database connections. See sqlite3_open_v2 for details.

The default open flags are SQLITE_OPEN_READ_WRITE | SQLITE_OPEN_CREATE | SQLITE_OPEN_URI | SQLITE_OPEN_NO_MUTEX. See Connection::open for some discussion about these flags.

Implementations

impl OpenFlags

pub const SQLITE_OPEN_READ_ONLY: Self = _

The database is opened in read-only mode. If the database does not already exist, an error is returned.

pub const SQLITE_OPEN_READ_WRITE: Self = _

The database is opened for reading and writing if possible, or reading only if the file is write protected by the operating system. In either case the database must already exist, otherwise an error is returned.

pub const SQLITE_OPEN_CREATE: Self = _

The database is created if it does not already exist

pub const SQLITE_OPEN_URI: Self = _

The filename can be interpreted as a URI if this flag is set.

pub const SQLITE_OPEN_MEMORY: Self = _

The database will be opened as an in-memory database.

pub const SQLITE_OPEN_NO_MUTEX: Self = _

The new database connection will not use a per-connection mutex (the connection will use the “multi-thread” threading mode, in SQLite parlance).

This is used by default, as proper Send/Sync usage (in particular, the fact that Connection does not implement Sync) ensures thread-safety without the need to perform locking around all calls.

pub const SQLITE_OPEN_FULL_MUTEX: Self = _

The new database connection will use a per-connection mutex – the “serialized” threading mode, in SQLite parlance.

Caveats

This flag should probably never be used with rusqlite, as we ensure thread-safety statically (we implement Send and not Sync). That said

Critically, even if this flag is used, the Connection is not safe to use across multiple threads simultaneously. To access a database from multiple threads, you should either create multiple connections, one for each thread (if you have very many threads, wrapping the rusqlite::Connection in a mutex is also reasonable).

This is both because of the additional per-connection state stored by rusqlite (for example, the prepared statement cache), and because not all of SQLites functions are fully thread safe, even in serialized/SQLITE_OPEN_FULLMUTEX mode.

All that said, it’s fairly harmless to enable this flag with rusqlite, it will just slow things down while providing no benefit.

pub const SQLITE_OPEN_SHARED_CACHE: Self = _

The database is opened with shared cache enabled.

This is frequently useful for in-memory connections, but note that broadly speaking it’s discouraged by SQLite itself, which states “Any use of shared cache is discouraged” in the official documentation.

pub const SQLITE_OPEN_PRIVATE_CACHE: Self = _

The database is opened shared cache disabled.

pub const SQLITE_OPEN_NOFOLLOW: Self = _

The database filename is not allowed to be a symbolic link. (3.31.0)

pub const SQLITE_OPEN_EXRESCODE: Self = _

Extended result codes. (3.37.0)

impl OpenFlags

pub const fn empty() -> Self

Returns an empty set of flags.

pub const fn all() -> Self

Returns the set containing all flags.

pub const fn bits(&self) -> c_int

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: c_int) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: c_int) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const fn from_bits_retain(bits: c_int) -> Self

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

pub fn from_name(name: &str) -> Option<Self>

Get the value for a flag from its stringified name.

Names are case-sensitive, so must correspond exactly to the identifier given to the flag.

pub const fn iter(&self) -> Iter

Iterate over enabled flag values.

pub const fn iter_names(&self) -> IterRaw

Iterate over enabled flag values with their stringified names.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Self)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Self, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the “bit-clear” operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> Self

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.