rusqlite

Struct Transaction

pub struct Transaction<'conn> {
    conn: &'conn Connection,
    drop_behavior: DropBehavior,
}

Represents a transaction on a database connection.

Note

Transactions will roll back by default. Use commit method to explicitly commit the transaction, or use set_drop_behavior to change what happens when the transaction is dropped.

Example

fn perform_queries(conn: &mut Connection) -> Result<()> {
    let tx = conn.transaction()?;

    do_queries_part_1(&tx)?; // tx causes rollback if this fails
    do_queries_part_2(&tx)?; // tx causes rollback if this fails

    tx.commit()
}

Fields
conn: &'conn Connectiondrop_behavior: DropBehavior
Implementations

impl Transaction<'_>

pub fn new( conn: &mut Connection, behavior: TransactionBehavior, ) -> Result<Transaction<'_>>

Begin a new transaction. Cannot be nested; see savepoint for nested transactions.

Even though we don’t mutate the connection, we take a &mut Connection to prevent nested transactions on the same connection. For cases where this is unacceptable, Transaction::new_unchecked is available.

pub fn new_unchecked( conn: &Connection, behavior: TransactionBehavior, ) -> Result<Transaction<'_>>

Begin a new transaction, failing if a transaction is open.

If a transaction is already open, this will return an error. Where possible, Transaction::new should be preferred, as it provides a compile-time guarantee that transactions are not nested.

pub fn savepoint(&mut self) -> Result<Savepoint<'_>>

Starts a new savepoint, allowing nested transactions.

Note

Just like outer level transactions, savepoint transactions rollback by default.

Example

fn perform_queries(conn: &mut Connection) -> Result<()> {
    let mut tx = conn.transaction()?;

    {
        let sp = tx.savepoint()?;
        if perform_queries_part_1_succeeds(&sp) {
            sp.commit()?;
        }
        // otherwise, sp will rollback
    }

    tx.commit()
}

pub fn savepoint_with_name<T: Into<String>>( &mut self, name: T, ) -> Result<Savepoint<'_>>

Create a new savepoint with a custom savepoint name. See savepoint().

pub fn drop_behavior(&self) -> DropBehavior

Get the current setting for what happens to the transaction when it is dropped.

pub fn set_drop_behavior(&mut self, drop_behavior: DropBehavior)

Configure the transaction to perform the specified action when it is dropped.

pub fn commit(self) -> Result<()>

A convenience method which consumes and commits a transaction.

fn commit_(&mut self) -> Result<()>

pub fn rollback(self) -> Result<()>

A convenience method which consumes and rolls back a transaction.

fn rollback_(&mut self) -> Result<()>

pub fn finish(self) -> Result<()>

Consumes the transaction, committing or rolling back according to the current setting (see drop_behavior).

Functionally equivalent to the Drop implementation, but allows callers to see any errors that occur.

fn finish_(&mut self) -> Result<()>