rusqlite

Trait BindIndex

pub trait BindIndex: Sealed {
    // Required method
    fn idx(&self, stmt: &Statement<'_>) -> Result<usize>;
}

A trait implemented by types that can index into parameters of a statement.

It is only implemented for usize and &str and &CStr.

Required Methods

fn idx(&self, stmt: &Statement<'_>) -> Result<usize>

Returns the index of the associated parameter, or Error if no such parameter exists.