Trait rusqlite::RowIndex

pub trait RowIndex {
    fn idx(&self, stmt: &Statement) -> Result<usize>;
}

A trait implemented by types that can index into columns of a row.

Required methods

fn idx(&self, stmt: &Statement) -> Result<usize>

Returns the index of the appropriate column, or None if no such column exists.

Loading content...