Enum rusqlite::types::ValueRef

pub enum ValueRef<'a> {
    Null,
    Integer(i64),
    Real(f64),
    Text(&'a str),
    Blob(&'a [u8]),
}

A non-owning dynamic type value. Typically the memory backing this value is owned by SQLite.

See Value for an owning dynamic type value.

Variants

Null

The value is a NULL value.

Integer

The value is a signed integer.

Real

The value is a floating point number.

Text

The value is a text string.

Blob

The value is a blob of data

Methods

impl<'a> ValueRef<'a>

fn as_i64(&self) -> Result<i64>

If self is case Integer, returns the integral value. Otherwise, returns Err(Error::InvalidColumnType).

fn as_f64(&self) -> Result<f64>

If self is case Real, returns the floating point value. Otherwise, returns Err(Error::InvalidColumnType).

fn as_str(&self) -> Result<&str>

If self is case Text, returns the string value. Otherwise, returns Err(Error::InvalidColumnType).

fn as_blob(&self) -> Result<&[u8]>

If self is case Blob, returns the byte slice. Otherwise, returns Err(Error::InvalidColumnType).