rusqlite

Module vtab

Create virtual tables.

Follow these steps to create your own virtual table:

Write implementation of VTab and VTabCursor traits.

Create an instance of the Module structure specialized for VTab impl. from step 1.

Register your Module structure using Connection::create_module.

Run a CREATE VIRTUAL TABLE command that specifies the new module in the USING clause.

(See SQLite doc)

Structs

Context

Context is used by VTabCursor::column to specify the cell value.

IndexConstraint

WHERE clause constraint.

IndexConstraintAndUsageIter

Iterate on index constraint and its associated usage.

IndexConstraintIter

feature = "vtab"

IndexConstraintUsage

Information about what parameters to pass to VTabCursor::filter.

IndexFlags

Virtual table scan flags See Function Flags for details.

IndexInfo

Pass information into and receive the reply from the VTab::best_index method.

Module

Virtual table module

OrderBy

A column of the ORDER BY clause.

OrderByIter

feature = "vtab"

VTabConnection

feature = "vtab"

ValueIter

Values iterator.

Values

Wrapper to VTabCursor::filter arguments, the values requested by VTab::best_index.

sqlite3_vtab

sqlite3_vtab_cursor

Enums

IndexConstraintOp

Index constraint operator. See Virtual Table Constraint Operator Codes for details.

VTabConfig

Virtual table configuration options

VTabKind

Virtual table kind

Traits

CreateVTab

Read-only virtual table instance trait.

TransactionVTab

Writable virtual table instance trait with transaction support trait.

UpdateVTab

Writable virtual table instance trait.

VTab

Eponymous-only virtual table instance trait.

VTabCursor

Virtual table cursor trait.

Functions

dequote

Dequote string

eponymous_only_module

Create an eponymous only virtual table implementation.

escape_double_quote

Escape double-quote (") character occurrences by doubling them ("").

parameter

<param_name>=['"]?<param_value>['"]? => (<param_name>, <param_value>)

parse_boolean

The boolean can be one of:

read_only_module

Create a read-only virtual table implementation.

update_module

Create a modifiable virtual table implementation.

update_module_with_tx

Create a modifiable virtual table implementation with support for transactions.