Enum rusqlite::config::DbConfig

#[non_exhaustive]

#[repr(i32)]
pub enum DbConfig {

    SQLITE_DBCONFIG_ENABLE_FKEY = 1_002,
    SQLITE_DBCONFIG_ENABLE_TRIGGER = 1_003,
    SQLITE_DBCONFIG_ENABLE_FTS3_TOKENIZER = 1_004,
    SQLITE_DBCONFIG_NO_CKPT_ON_CLOSE = 1_006,
    SQLITE_DBCONFIG_ENABLE_QPSG = 1_007,
    SQLITE_DBCONFIG_TRIGGER_EQP = 1_008,
    SQLITE_DBCONFIG_RESET_DATABASE = 1_009,
    SQLITE_DBCONFIG_DEFENSIVE = 1_010,
    SQLITE_DBCONFIG_WRITABLE_SCHEMA = 1_011,
    SQLITE_DBCONFIG_LEGACY_ALTER_TABLE = 1_012,
    SQLITE_DBCONFIG_DQS_DML = 1_013,
    SQLITE_DBCONFIG_DQS_DDL = 1_014,
    SQLITE_DBCONFIG_ENABLE_VIEW = 1_015,
    SQLITE_DBCONFIG_LEGACY_FILE_FORMAT = 1_016,
    SQLITE_DBCONFIG_TRUSTED_SCHEMA = 1_017,

}

Database Connection Configuration Options See Database Connection Configuration Options for details.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

SQLITE_DBCONFIG_ENABLE_FKEY = 1_002

Enable or disable the enforcement of foreign key constraints.

SQLITE_DBCONFIG_ENABLE_TRIGGER = 1_003

Enable or disable triggers.

SQLITE_DBCONFIG_ENABLE_FTS3_TOKENIZER = 1_004

Enable or disable the fts3_tokenizer() function which is part of the FTS3 full-text search engine extension.

SQLITE_DBCONFIG_NO_CKPT_ON_CLOSE = 1_006

In WAL mode, enable or disable the checkpoint operation before closing the connection.

SQLITE_DBCONFIG_ENABLE_QPSG = 1_007

Activates or deactivates the query planner stability guarantee (QPSG).

SQLITE_DBCONFIG_TRIGGER_EQP = 1_008

Includes or excludes output for any operations performed by trigger programs from the output of EXPLAIN QUERY PLAN commands.

SQLITE_DBCONFIG_RESET_DATABASE = 1_009

Activates or deactivates the “reset” flag for a database connection. Run VACUUM with this flag set to reset the database.

SQLITE_DBCONFIG_DEFENSIVE = 1_010

Activates or deactivates the “defensive” flag for a database connection.

SQLITE_DBCONFIG_WRITABLE_SCHEMA = 1_011

Activates or deactivates the “writable_schema” flag.

SQLITE_DBCONFIG_LEGACY_ALTER_TABLE = 1_012

Activates or deactivates the legacy behavior of the ALTER TABLE RENAME command.

SQLITE_DBCONFIG_DQS_DML = 1_013

Activates or deactivates the legacy double-quoted string literal misfeature for DML statements only.

SQLITE_DBCONFIG_DQS_DDL = 1_014

Activates or deactivates the legacy double-quoted string literal misfeature for DDL statements.

SQLITE_DBCONFIG_ENABLE_VIEW = 1_015

Enable or disable views.

SQLITE_DBCONFIG_LEGACY_FILE_FORMAT = 1_016

Activates or deactivates the legacy file format flag.

SQLITE_DBCONFIG_TRUSTED_SCHEMA = 1_017

Tells SQLite to assume that database schemas (the contents of the sqlite_master tables) are untainted by malicious content.