Enum rusqlite::config::DbConfig

#[repr(i32)]
#[non_exhaustive]
pub enum DbConfig {

    SQLITE_DBCONFIG_ENABLE_FKEY,
    SQLITE_DBCONFIG_ENABLE_TRIGGER,
    SQLITE_DBCONFIG_ENABLE_FTS3_TOKENIZER,
    SQLITE_DBCONFIG_NO_CKPT_ON_CLOSE,
    SQLITE_DBCONFIG_ENABLE_QPSG,
    SQLITE_DBCONFIG_TRIGGER_EQP,
    SQLITE_DBCONFIG_DEFENSIVE,
    SQLITE_DBCONFIG_WRITABLE_SCHEMA,
    SQLITE_DBCONFIG_LEGACY_ALTER_TABLE,
    SQLITE_DBCONFIG_DQS_DML,
    SQLITE_DBCONFIG_DQS_DDL,
    SQLITE_DBCONFIG_ENABLE_VIEW,
    SQLITE_DBCONFIG_LEGACY_FILE_FORMAT,
    SQLITE_DBCONFIG_TRUSTED_SCHEMA,

}

Database Connection Configuration Options See Database Connection Configuration Options for details.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

SQLITE_DBCONFIG_ENABLE_FKEY

Enable or disable the enforcement of foreign key constraints.

SQLITE_DBCONFIG_ENABLE_TRIGGER

Enable or disable triggers.

SQLITE_DBCONFIG_ENABLE_FTS3_TOKENIZER

Enable or disable the fts3_tokenizer() function which is part of the FTS3 full-text search engine extension.

SQLITE_DBCONFIG_NO_CKPT_ON_CLOSE

In WAL mode, enable or disable the checkpoint operation before closing the connection.

SQLITE_DBCONFIG_ENABLE_QPSG

Activates or deactivates the query planner stability guarantee (QPSG).

SQLITE_DBCONFIG_TRIGGER_EQP

Includes or excludes output for any operations performed by trigger programs from the output of EXPLAIN QUERY PLAN commands.

SQLITE_DBCONFIG_DEFENSIVE

Activates or deactivates the “defensive” flag for a database connection.

SQLITE_DBCONFIG_WRITABLE_SCHEMA

Activates or deactivates the “writable_schema” flag.

SQLITE_DBCONFIG_LEGACY_ALTER_TABLE

Activates or deactivates the legacy behavior of the ALTER TABLE RENAME command.

SQLITE_DBCONFIG_DQS_DML

Activates or deactivates the legacy double-quoted string literal misfeature for DML statements only.

SQLITE_DBCONFIG_DQS_DDL

Activates or deactivates the legacy double-quoted string literal misfeature for DDL statements.

SQLITE_DBCONFIG_ENABLE_VIEW

Enable or disable views.

SQLITE_DBCONFIG_LEGACY_FILE_FORMAT

Activates or deactivates the legacy file format flag.

SQLITE_DBCONFIG_TRUSTED_SCHEMA

Tells SQLite to assume that database schemas (the contents of the sqlite_master tables) are untainted by malicious content.