rusqlite

Enum Error

#[non_exhaustive]
pub enum Error {

    SqliteFailure(Error, Option<String>),
    SqliteSingleThreadedMode,
    FromSqlConversionFailure(usize, Type, Box<dyn Error + Send + Sync + 'static>),
    IntegralValueOutOfRange(usize, i64),
    Utf8Error(Utf8Error),
    NulError(NulError),
    InvalidParameterName(String),
    InvalidPath(PathBuf),
    ExecuteReturnedResults,
    QueryReturnedNoRows,
    InvalidColumnIndex(usize),
    InvalidColumnName(String),
    InvalidColumnType(usize, String, Type),
    StatementChangedRows(usize),
    InvalidFunctionParameterType(usize, Type),
    InvalidFilterParameterType(usize, Type),
    UserFunctionError(Box<dyn Error + Send + Sync + 'static>),
    ToSqlConversionFailure(Box<dyn Error + Send + Sync + 'static>),
    InvalidQuery,
    ModuleError(String),
    UnwindingPanic,
    GetAuxWrongType,
    MultipleStatement,
    InvalidParameterCount(usize, usize),
    SqlInputError {
        error: Error,
        msg: String,
        sql: String,
        offset: c_int,
    },
    InvalidDatabaseIndex(usize),

}

Enum listing possible errors from rusqlite.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

SqliteFailure(Error, Option<String>)

An error from an underlying SQLite call.

SqliteSingleThreadedMode

Error reported when attempting to open a connection when SQLite was configured to allow single-threaded use only.

FromSqlConversionFailure(usize, Type, Box<dyn Error + Send + Sync + 'static>)

Error when the value of a particular column is requested, but it cannot be converted to the requested Rust type.

IntegralValueOutOfRange(usize, i64)

Error when SQLite gives us an integral value outside the range of the requested type (e.g., trying to get the value 1000 into a u8). The associated usize is the column index, and the associated i64 is the value returned by SQLite.

Utf8Error(Utf8Error)

Error converting a string to UTF-8.

NulError(NulError)

Error converting a string to a C-compatible string because it contained an embedded nul.

InvalidParameterName(String)

Error when using SQL named parameters and passing a parameter name not present in the SQL.

InvalidPath(PathBuf)

Error converting a file path to a string.

ExecuteReturnedResults

Error returned when an execute call returns rows.

QueryReturnedNoRows

Error when a query that was expected to return at least one row (e.g., for query_row) did not return any.

InvalidColumnIndex(usize)

Error when the value of a particular column is requested, but the index is out of range for the statement.

InvalidColumnName(String)

Error when the value of a named column is requested, but no column matches the name for the statement.

InvalidColumnType(usize, String, Type)

Error when the value of a particular column is requested, but the type of the result in that column cannot be converted to the requested Rust type.

StatementChangedRows(usize)

Error when a query that was expected to insert one row did not insert any or insert many.

InvalidFunctionParameterType(usize, Type)

Error returned by functions::Context::get when the function argument cannot be converted to the requested type.

InvalidFilterParameterType(usize, Type)

Error returned by vtab::Values::get when the filter argument cannot be converted to the requested type.

UserFunctionError(Box<dyn Error + Send + Sync + 'static>)

An error case available for implementors of custom user functions (e.g., create_scalar_function).

ToSqlConversionFailure(Box<dyn Error + Send + Sync + 'static>)

Error available for the implementors of the ToSql trait.

InvalidQuery

Error when the SQL is not a SELECT, is not read-only.

ModuleError(String)

An error case available for implementors of custom modules (e.g., create_module).

UnwindingPanic

An unwinding panic occurs in a UDF (user-defined function).

GetAuxWrongType

An error returned when Context::get_aux attempts to retrieve data of a different type than what had been stored using Context::set_aux.

MultipleStatement

Error when the SQL contains multiple statements.

InvalidParameterCount(usize, usize)

Error when the number of bound parameters does not match the number of parameters in the query. The first usize is how many parameters were given, the 2nd is how many were expected.

SqlInputError

Error referencing a specific token in the input SQL

Fields

error: Error

error code

msg: String

error message

sql: String

SQL input

offset: c_int

byte offset of the start of invalid token

InvalidDatabaseIndex(usize)

Error when the schema of a particular database is requested, but the index is out of range.

Implementations

impl Error

pub fn sqlite_error(&self) -> Option<&Error>

Returns the underlying SQLite error if this is Error::SqliteFailure.

pub fn sqlite_error_code(&self) -> Option<ErrorCode>

Returns the underlying SQLite error code if this is Error::SqliteFailure.