Struct rusqlite::CachedStatement

pub struct CachedStatement<'conn> { /* private fields */ }

Cacheable statement.

Statement will return automatically to the cache by default. If you want the statement to be discarded, call discard() on it.

Implementations

impl CachedStatement<'_>

pub fn discard(self)

Discard the statement, preventing it from being returned to its Connection’s collection of cached statements.