Struct rusqlite::Row

pub struct Row<'a, 'stmt> { /* fields omitted */ }

A single result row of a query.

Methods

impl<'a, 'stmt> Row<'a, 'stmt>

pub fn get<I: RowIndex, T: FromSql>(&self, idx: I) -> T

Get the value of a particular column of the result row.

Failure

Panics if calling row.get_checked(idx) would return an error, including:

If the underlying SQLite column type is not a valid type as a source for T

If the underlying SQLite integral value is outside the range representable by T

If idx is outside the range of columns in the returned query

pub fn get_checked<I: RowIndex, T: FromSql>(&self, idx: I) -> Result<T>

Get the value of a particular column of the result row.

Failure

Returns an Error::InvalidColumnType if the underlying SQLite column type is not a valid type as a source for T.

Returns an Error::InvalidColumnIndex if idx is outside the valid column range for this row.

Returns an Error::InvalidColumnName if idx is not a valid column name for this row.

pub fn column_count(&self) -> i32

Return the number of columns in the current row.