Struct rusqlite::Row

pub struct Row<'stmt> { /* private fields */ }

A single result row of a query.

Implementations

impl<'stmt> Row<'stmt>

pub fn column_names(&self) -> Vec<&str>

Get all the column names of the Row.

pub fn column_count(&self) -> usize

Return the number of columns in the current row.

pub fn column_name(&self, col: usize) -> Result<&str>

Return the name of the column.

pub fn column_index(&self, name: &str) -> Result<usize>

Return the index of the column.

pub fn columns(&self) -> Vec<Column<'_>>

Returns a slice describing the columns of the Row.

impl<'stmt> Row<'stmt>

pub fn get_unwrap<I: RowIndex, T: FromSql>(&self, idx: I) -> T

Get the value of a particular column of the result row.

Failure

Panics if calling row.get(idx) would return an error, including:

If the underlying SQLite column type is not a valid type as a source for T

If the underlying SQLite integral value is outside the range representable by T

If idx is outside the range of columns in the returned query

pub fn get<I: RowIndex, T: FromSql>(&self, idx: I) -> Result<T>

Get the value of a particular column of the result row.

Failure

Returns an Error::InvalidColumnType if the underlying SQLite column type is not a valid type as a source for T.

Returns an Error::InvalidColumnIndex if idx is outside the valid column range for this row.

Returns an Error::InvalidColumnName if idx is not a valid column name for this row.

If the result type is i128 (which requires the i128_blob feature to be enabled), and the underlying SQLite column is a blob whose size is not 16 bytes, Error::InvalidColumnType will also be returned.

pub fn get_ref<I: RowIndex>(&self, idx: I) -> Result<ValueRef<'_>>

Get the value of a particular column of the result row as a ValueRef, allowing data to be read out of a row without copying.

This ValueRef is valid only as long as this Row, which is enforced by it’s lifetime. This means that while this method is completely safe, it can be somewhat difficult to use, and most callers will be better served by get or get_unwrap.

Failure

Returns an Error::InvalidColumnIndex if idx is outside the valid column range for this row.

Returns an Error::InvalidColumnName if idx is not a valid column name for this row.

pub fn get_ref_unwrap<I: RowIndex>(&self, idx: I) -> ValueRef<'_>

Get the value of a particular column of the result row as a ValueRef, allowing data to be read out of a row without copying.

This ValueRef is valid only as long as this Row, which is enforced by it’s lifetime. This means that while this method is completely safe, it can be difficult to use, and most callers will be better served by get or get_unwrap.

Failure

Panics if calling row.get_ref(idx) would return an error, including:

If idx is outside the range of columns in the returned query.

If idx is not a valid column name for this row.

pub fn get_raw_checked<I: RowIndex>(&self, idx: I) -> Result<ValueRef<'_>>

👎Deprecated: Use get_ref instead.

Renamed to get_ref.

pub fn get_raw<I: RowIndex>(&self, idx: I) -> ValueRef<'_>

👎Deprecated: Use get_ref_unwrap instead.

Renamed to get_ref_unwrap.