Struct rusqlite::Rows

pub struct Rows<'stmt> { /* private fields */ }

An handle for the resulting rows of a query.

Implementations

impl<'stmt> Rows<'stmt>

pub fn column_names(&self) -> Option<Vec<&str>>

Get all the column names.

pub fn column_count(&self) -> Option<usize>

Return the number of columns.

pub fn column_name(&self, col: usize) -> Option<Result<&str>>

Return the name of the column.

pub fn column_index(&self, name: &str) -> Option<Result<usize>>

Return the index of the column.

pub fn columns(&self) -> Option<Vec<Column<'_>>>

Returns a slice describing the columns of the Rows.

impl<'stmt> Rows<'stmt>

pub fn next(&mut self) -> Result<Option<&Row<'stmt>>>

Attempt to get the next row from the query. Returns Ok(Some(Row)) if there is another row, Err(...) if there was an error getting the next row, and Ok(None) if all rows have been retrieved.

Note

This interface is not compatible with Rust’s Iterator trait, because the lifetime of the returned row is tied to the lifetime of self. This is a fallible “streaming iterator”. For a more natural interface, consider using query_map or query_and_then instead, which return types that implement Iterator.

pub fn map<F, B>(self, f: F) -> Map<'stmt, F>where
F: FnMut(&Row<'_>) -> Result<B>,

Map over this Rows, converting it to a Map, which implements FallibleIterator.

use fallible_iterator::FallibleIterator;
fn query(stmt: &mut Statement) -> Result<Vec<i64>> {
    let rows = stmt.query([])?;
    rows.map(|r| r.get(0)).collect()
}

pub fn mapped<F, B>(self, f: F) -> MappedRows<'stmt, F> where
F: FnMut(&Row<'_>) -> Result<B>,

Map over this Rows, converting it to a MappedRows, which implements Iterator.

pub fn and_then<F, T, E>(self, f: F) -> AndThenRows<'stmt, F> where
F: FnMut(&Row<'_>) -> Result<T, E>,

Map over this Rows with a fallible function, converting it to a AndThenRows, which implements Iterator (instead of FallibleStreamingIterator).