rusqlite

Struct Rows

pub struct Rows<'stmt> { /* private fields */ }

A handle (lazy fallible streaming iterator) for the resulting rows of a query.

Implementations

impl<'stmt> Rows<'stmt>

pub fn next(&mut self) -> Result<Option<&Row<'stmt>>>

Attempt to get the next row from the query. Returns Ok(Some(Row)) if there is another row, Err(...) if there was an error getting the next row, and Ok(None) if all rows have been retrieved.

Note

This interface is not compatible with Rust’s Iterator trait, because the lifetime of the returned row is tied to the lifetime of self. This is a fallible “streaming iterator”. For a more natural interface, consider using query_map or query_and_then instead, which return types that implement Iterator.

pub fn map<F, B>(self, f: F) -> Map<'stmt, F>
where F: FnMut(&Row<'_>) -> Result<B>,

Map over this Rows, converting it to a Map, which implements FallibleIterator.

use fallible_iterator::FallibleIterator;
fn query(stmt: &mut Statement) -> Result<Vec<i64>> {
    let rows = stmt.query([])?;
    rows.map(|r| r.get(0)).collect()
}

pub fn mapped<F, B>(self, f: F) -> MappedRows<'stmt, F>
where F: FnMut(&Row<'_>) -> Result<B>,

Map over this Rows, converting it to a MappedRows, which implements Iterator.

pub fn and_then<F, T, E>(self, f: F) -> AndThenRows<'stmt, F>
where F: FnMut(&Row<'_>) -> Result<T, E>,

Map over this Rows with a fallible function, converting it to a AndThenRows, which implements Iterator (instead of FallibleStreamingIterator).

pub fn as_ref(&self) -> Option<&Statement<'stmt>>

Give access to the underlying statement