Struct rusqlite::Savepoint

pub struct Savepoint<'conn> { /* private fields */ }

Represents a savepoint on a database connection.

Note

Savepoints will roll back by default. Use commit method to explicitly commit the savepoint, or use set_drop_behavior to change what happens when the savepoint is dropped.

Example

fn perform_queries(conn: &mut Connection) -> Result<()> {
    let sp = conn.savepoint()?;

    do_queries_part_1(&sp)?; // sp causes rollback if this fails
    do_queries_part_2(&sp)?; // sp causes rollback if this fails

    sp.commit()
}

Implementations

impl Savepoint<'_>

pub fn new(conn: &mut Connection) -> Result<Savepoint<'_>>

Begin a new savepoint. Can be nested.

pub fn with_name<T: Into<String>>( conn: &mut Connection, name: T ) -> Result<Savepoint<'_>>

Begin a new savepoint with a user-provided savepoint name.

pub fn savepoint(&mut self) -> Result<Savepoint<'_>>

Begin a nested savepoint.

pub fn savepoint_with_name<T: Into<String>>( &mut self, name: T ) -> Result<Savepoint<'_>>

Begin a nested savepoint with a user-provided savepoint name.

pub fn drop_behavior(&self) -> DropBehavior

Get the current setting for what happens to the savepoint when it is dropped.

pub fn set_drop_behavior(&mut self, drop_behavior: DropBehavior)

Configure the savepoint to perform the specified action when it is dropped.

pub fn commit(self) -> Result<()>

A convenience method which consumes and commits a savepoint.

pub fn rollback(&mut self) -> Result<()>

A convenience method which rolls back a savepoint.

Note

Unlike Transactions, savepoints remain active after they have been rolled back, and can be rolled back again or committed.

pub fn finish(self) -> Result<()>

Consumes the savepoint, committing or rolling back according to the current setting (see drop_behavior).

Functionally equivalent to the Drop implementation, but allows callers to see any errors that occur.