Struct rusqlite::Transaction

pub struct Transaction<'conn> { /* fields omitted */ }

Represents a transaction on a database connection.

Note

Transactions will roll back by default. Use commit method to explicitly commit the transaction, or use set_drop_behavior to change what happens when the transaction is dropped.

Example

fn perform_queries(conn: &mut Connection) -> Result<()> {
    let tx = try!(conn.transaction());

    try!(do_queries_part_1(&tx)); // tx causes rollback if this fails
    try!(do_queries_part_2(&tx)); // tx causes rollback if this fails

    tx.commit()
}

Methods

impl<'conn> Transaction<'conn>

pub fn new(
conn: &mut Connection,
behavior: TransactionBehavior
) -> Result<Transaction>

Begin a new transaction. Cannot be nested; see savepoint for nested transactions.

pub fn savepoint(&mut self) -> Result<Savepoint>

Starts a new savepoint, allowing nested transactions.

Note

Just like outer level transactions, savepoint transactions rollback by default.

Example

fn perform_queries(conn: &mut Connection) -> Result<()> {
    let mut tx = try!(conn.transaction());

    {
        let sp = try!(tx.savepoint());
        if perform_queries_part_1_succeeds(&sp) {
            try!(sp.commit());
        }
        // otherwise, sp will rollback
    }

    tx.commit()
}

pub fn savepoint_with_name<T: Into<String>>(
&mut self,
name: T
) -> Result<Savepoint>

Create a new savepoint with a custom savepoint name. See savepoint().

pub fn drop_behavior(&self) -> DropBehavior

Get the current setting for what happens to the transaction when it is dropped.

pub fn set_drop_behavior(&mut self, drop_behavior: DropBehavior)

Configure the transaction to perform the specified action when it is dropped.

pub fn commit(self) -> Result<()>

A convenience method which consumes and commits a transaction.

pub fn rollback(self) -> Result<()>

A convenience method which consumes and rolls back a transaction.

pub fn finish(self) -> Result<()>

Consumes the transaction, committing or rolling back according to the current setting (see drop_behavior).

Functionally equivalent to the Drop implementation, but allows callers to see any errors that occur.