Trait rusqlite::OptionalExtension

pub trait OptionalExtension<T> {
    fn optional(self) -> Result<Option<T>>;
}

See the method documentation.

Required Methods

fn optional(self) -> Result<Option<T>>

Converts a Result<T> into a Result<Option<T>>.

By default, Rusqlite treats 0 rows being returned from a query that is expected to return 1 row as an error. This method will handle that error, and give you back an Option<T> instead.