Module rusqlite::types

Traits dealing with SQLite data types.

SQLite uses a dynamic type system. Implementations of the ToSql and FromSql traits are provided for the basic types that SQLite provides methods for:

Integers (i32 and i64; SQLite uses i64 internally, so getting an i32 will truncate if the value is too large or too small).

Reals (f64)

Strings (String and &str)

Blobs (Vec<u8> and &[u8])

Additionally, if the time feature is enabled, implementations are provided for time::OffsetDateTime that use the RFC 3339 date/time format, "%Y-%m-%dT%H:%M:%S.%fZ", to store time values as strings. These values can be parsed by SQLite's builtin datetime functions. If you want different storage for datetimes, you can use a newtype.

ToSql and FromSql are also implemented for Option<T> where T implements ToSql or FromSql for the cases where you want to know if a value was NULL (which gets translated to None).

Structs

Null
Empty struct that can be used to fill in a query parameter as NULL.

Enums

FromSqlError
Enum listing possible errors from FromSql trait.

ToSqlOutput
ToSqlOutput represents the possible output types for implementors of the ToSql trait.

Type
SQLite data types. See Fundamental Datatypes.

Value
Owning dynamic type value. Value's type is typically dictated by SQLite (not by the caller).

ValueRef
A non-owning dynamic type value. Typically the memory backing this value is owned by SQLite.

Traits

FromSql
A trait for types that can be created from a SQLite value.

ToSql
A trait for types that can be converted into SQLite values.

Type Definitions

FromSqlResult
Result type for implementors of the FromSql trait.