Enum rusqlite::types::ValueRef

pub enum ValueRef<'a> {
    Null,
    Integer(i64),
    Real(f64),
    Text(&'a str),
    Blob(&'a [u8]),
}

A non-owning dynamic type value. Typically the memory backing this value is owned by SQLite.

See Value for an owning dynamic type value.

Variants
Null

The value is a NULL value.

Integer(i64)

The value is a signed integer.

Real(f64)

The value is a floating point number.

Text(&'a str)

The value is a text string.

Blob(&'a [u8])

The value is a blob of data

Methods

impl<'a> ValueRef<'a>

pub fn data_type(&self) -> Type

impl<'a> ValueRef<'a>

pub fn as_i64(&self) -> FromSqlResult<i64>

If self is case Integer, returns the integral value. Otherwise, returns Err(Error::InvalidColumnType).

pub fn as_f64(&self) -> FromSqlResult<f64>

If self is case Real, returns the floating point value. Otherwise, returns Err(Error::InvalidColumnType).

pub fn as_str(&self) -> FromSqlResult<&str>

If self is case Text, returns the string value. Otherwise, returns Err(Error::InvalidColumnType).

pub fn as_blob(&self) -> FromSqlResult<&[u8]>

If self is case Blob, returns the byte slice. Otherwise, returns Err(Error::InvalidColumnType).